Work, stores, items, inventory, skill checks and some ui changes

+++ Assets/Scripts/Interaction.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
@@ -15,8 +16,10 @@ public class Interaction : ScriptableObject
     public int[] requiredLevels;
     public bool skillIncrease = false;
     public string[] skillsToChange;
-    public int[] skillLevelstoChange;
+    public float[] skillLevelstoChange;
     public Thought[] induceThoughts;
     // Do player's needs decay while perforing the action?
     public bool needsDecay = true;
 }
+
+
